refactor: split code review integration tests into separate files by endpoint
